requirements.txt
Flask
flask-cors
pymongo
python-dotenv
werkzeug

